% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_monaR-package.R
\docType{data}
\name{example_id_query}
\alias{example_id_query}
\title{Example mona_query results}
\format{
tibble with 181 rows and 13 variables
}
\usage{
example_id_query
}
\description{
An example result generated from mona_query('atrazine', 'name')
}
\keyword{datasets}
