% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_monaR-package.R
\docType{data}
\name{example_spec_query}
\alias{example_spec_query}
\title{Example mona_querySpec results}
\format{
tibble with 15 rows and 2 variables
}
\usage{
example_spec_query
}
\description{
An example result generated from
mona_querySpec(
   spectrum = atrazine_ms2,
   minSimilarity = 800,
   precursorMZ = 216,
   precursorToleranceDa =1)
}
\keyword{datasets}
