% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_query_by_mol_identifier.R
\name{get_tags}
\alias{get_tags}
\title{Get tags for query}
\usage{
get_tags(ionization = NULL, ms_level = NULL, source_introduction = NULL)
}
\arguments{
\item{ionization}{The ionization mode (positive, negative, both)}

\item{ms_level}{Which order of MS should be searched (all, MS1, MS2, MS3,
MS4).}

\item{source_introduction}{Which instrument platforms should be considered
(all, LC-MS, GC-MS, CE-MS)}
}
\value{
A named list of tags to add to a query
}
\description{
Get tags for query
}
