% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{mona_extract}
\alias{mona_extract}
\alias{mona_extract.mona_id_query}
\alias{mona_extract.mona_spec_query}
\title{Extract elements from a tibble returned by mona_query}
\usage{
mona_extract(df, what)

\method{mona_extract}{mona_id_query}(
  df,
  what = c("compound", "id", "dateCreated", "lastUpdated", "lastCurated", "metaData",
    "score", "spectrum", "tags", "library")
)

\method{mona_extract}{mona_spec_query}(df, what)
}
\arguments{
\item{df}{A tibble returned by mona_query or mona_querySpectrum}

\item{what}{Which element to extract}
}
\value{

}
\description{
Extract elements from a tibble returned by mona_query
}
\section{Methods (by class)}{
\itemize{
\item \code{mona_id_query}: Extract identifiers from ID queries

\item \code{mona_spec_query}: Extract identifiers from spectrum queries
}}

