% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_data.R
\name{mona_getChem}
\alias{mona_getChem}
\alias{mona_getChem.mona_id_query}
\alias{mona_getChem.mona_spec_query}
\title{Extract chemical data from MONA queries}
\usage{
mona_getChem(df, var)

\method{mona_getChem}{mona_id_query}(
  df,
  var = c("inchi", "inchiKey", "molFile", "names", "SMILES", "compound class",
    "molecular formula", "external id", "computed")
)

\method{mona_getChem}{mona_spec_query}(
  df,
  var = c("inchi", "inchiKey", "molFile", "names", "SMILES", "compound class",
    "molecular formula", "external id", "computed")
)
}
\arguments{
\item{df}{A tibble returned by a MoNA query}

\item{var}{The variable to return from the meta data}
}
\value{
a tibble containing extracted chemical data
}
\description{
Extract chemical data from MONA queries
}
\section{Methods (by class)}{
\itemize{
\item \code{mona_id_query}: Get values from meta data from ID query

\item \code{mona_spec_query}: Extract meta data from spectrum queries
}}

\examples{
mona_getChem(example_id_query, var = 'inchi')
mona_getChem(example_id_query, var = 'inchiKey')
mona_getChem(example_id_query, var = 'external id')

}
