% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_data.R
\name{mona_getMeta}
\alias{mona_getMeta}
\alias{mona_getMeta.mona_id_query}
\alias{mona_getMeta.mona_spec_query}
\title{Extract values from meta data records retrieved from MONA queries}
\usage{
mona_getMeta(df, var, value)

\method{mona_getMeta}{mona_id_query}(
  df,
  var = c("category", "name"),
  value = c("mass spectrometry", "chromatography", "focused ion", "accession", "date",
    "author", "license", "copyright", "publication", "exact mass", "instrument",
    "instrument type", "ms level", "ionization", "column", "column temperature",
    "injection temperature", "retention time", "precursor m/z", "precursor type",
    "mass accuracy", "mass error")
)

\method{mona_getMeta}{mona_spec_query}(
  df,
  var = c("category", "name", "submitter", "splash", "submitter", "library", "score"),
  value = c("mass spectrometry", "chromatography", "focused ion", "accession", "date",
    "author", "license", "copyright", "publication", "exact mass", "instrument",
    "instrument type", "ms level", "ionization", "column", "column temperature",
    "injection temperature", "retention time", "precursor m/z", "precursor type",
    "mass accuracy", "mass error")
)
}
\value{
A tibble with columns id, value
}
\description{
Extract values from meta data records retrieved from MONA queries
}
\section{Methods (by class)}{
\itemize{
\item \code{mona_id_query}: Get values from meta data from ID query

\item \code{mona_spec_query}: Extract meta data from spectrum query
}}

\examples{
mona_getMeta(example_id_query, var = 'category', value = 'mass spectrometry')
mona_getMeta(example_id_query, var = 'category', value = 'focused ion')
mona_getMeta(example_id_query, var = 'category', value = 'chromatography')
mona_getMeta(example_id_query, var = 'name', value = 'ionization')
mona_getMeta(example_id_query, var = 'name', value = 'ms level')
mona_getMeta(example_id_query, var = 'name', value = 'mass error')
mona_getMeta(example_spec_query, var = 'category', value = 'mass spectrometry')
mona_getMeta(example_spec_query, var = 'category', value = 'chromatography')
mona_getMeta(example_spec_query, var = 'name', value = 'mass error')

}
