% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_handling.R
\name{mona_getSpec}
\alias{mona_getSpec}
\alias{mona_getSpec.mona_id_query}
\alias{mona_getSpec.mona_spec_query}
\title{Extract spectra from MONA query results}
\usage{
mona_getSpec(df, ann = FALSE)

\method{mona_getSpec}{mona_id_query}(df, ann = FALSE)

\method{mona_getSpec}{mona_spec_query}(df, ann = FALSE)
}
\arguments{
\item{df}{A tibble generated by a mona_query or mona_querySpec call}

\item{ann}{T/F should annotations be merged with the spectra}
}
\value{
A list of spectra named with the spectrum id
}
\description{
Extract spectra from MONA query results
}
\section{Methods (by class)}{
\itemize{
\item \code{mona_id_query}: Extract spectrum from ID query

\item \code{mona_spec_query}: Extract spectra from spectrum query
}}

\examples{
mona_getSpec(dplyr::slice(example_id_query,1))
mona_getSpec(dplyr::slice(example_id_query,150), ann = TRUE)

mona_getSpec(dplyr::slice(example_spec_query, 1))
mona_getSpec(dplyr::filter(example_spec_query, example_spec_query$hit$id == 'SM841401'))
mona_getSpec(dplyr::filter(example_spec_query,example_spec_query$score > 0.8))
mona_getSpec(dplyr::slice(example_spec_query, 1), ann = TRUE)

}
