% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mona_query.R
\name{mona_parseSpec}
\alias{mona_parseSpec}
\alias{mona_parseSpec.character}
\alias{mona_parseSpec.data.frame}
\title{Interchange between mass spectrum formats for MONA queries and results}
\usage{
mona_parseSpec(spec)

\method{mona_parseSpec}{character}(spec)

\method{mona_parseSpec}{data.frame}(spec)
}
\arguments{
\item{spec}{A spectrum as either a character string of space deliminated
mz:intensity pairs or a dataframe with variables mz and intensity}
}
\value{
Provided a spectrum as a string returns a tibble and vice versa.
}
\description{
Interchange between mass spectrum formats for MONA queries and results
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Method for character strings

\item \code{data.frame}: Method for data frames
}}

\examples{

mona_parseSpec("70.065:16.558562 84.0443:6.200376 86.0963:16.435100
93.0697:3.403519 103.0541:4.278218 105.0697:10.396751 107.0854:23.813899")

}
