% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mona_query.R
\name{mona_query}
\alias{mona_query}
\title{Query Mass Bank of North America API}
\usage{
mona_query(
  query,
  from = c("text", "name", "InChIkey", "molform"),
  mass_tol_Da = 0.1,
  ionization = NULL
)
}
\arguments{
\item{query}{The query string}

\item{from}{The query type (text, name, InChIKey or molform)}
}
\value{
A tibble containing the search results or NA if no matches.
}
\description{
Submit a query to MONA API (https://mona.fiehnlab.ucdavis.edu) by text
string, molecular formula or (partial) InChI Key.
}
