% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_query_by_spectrum.R
\name{mona_querySpec}
\alias{mona_querySpec}
\title{Query MONA API by a mass spectrum}
\usage{
mona_querySpec(
  spectrum = NULL,
  minSimilarity = NULL,
  precursorMZ = NULL,
  precursorToleranceDa = NULL,
  precursorTolerancePPM = NULL
)
}
\arguments{
\item{spectrum}{A spectrum as either a dataframe with variables mz,intensity
or a space deliminated stinrg of mz:inensity pairs.}

\item{minSimilarity}{Minimum spectral similarity for searching}

\item{precursorMZ}{The precursor ion m/z}

\item{precursorToleranceDa}{Search tolerance in dalton}

\item{precursorTolerancePPM}{Search tolerance in ppm}
}
\value{
a tibble containing query results
}
\description{
Query MONA API by a mass spectrum
}
\examples{
 
   mona_querySpec(
       spectrum = atrazine_ms2, minSimilarity = 800,
       precursorMZ = 216, precursorToleranceDa = 1 
       )
}
