% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_print_methods.R
\name{print.mona_meta}
\alias{print.mona_meta}
\title{print method for meta data}
\usage{
\method{print}{mona_meta}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints tibble info from a meta data query
}
\description{
print method for meta data
}
\examples{
print(mona_getMeta(example_id_query, 'category', 'mass spectrometry'))
}
