% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_print_methods.R
\name{print.mona_spec_query}
\alias{print.mona_spec_query}
\title{print method for spec query}
\usage{
\method{print}{mona_spec_query}(x, ...)
}
\description{
print method for spec query
}
\examples{
print(example_spec_query)
}
