% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mona_query.R
\name{mona_identifiers}
\alias{mona_identifiers}
\title{Download the spectra identifier table from MONA}
\usage{
mona_identifiers(update = F)
}
\arguments{
\item{update}{(T/F) should the file be downloaded}
}
\value{
Saves table as an R-object
}
\description{
Download the spectra identifier table from MONA
}
