% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mona_query.R
\name{mona_parseSpec}
\alias{mona_parseSpec}
\alias{mona_parseSpec.character}
\alias{mona_parseSpec.data.frame}
\alias{mona_parseSpec.matrix}
\alias{mona_parseSpec.Spectrum2}
\title{Interchange between mass spectrum formats for MONA queries and results}
\usage{
mona_parseSpec(spec)

\method{mona_parseSpec}{character}(spec)

\method{mona_parseSpec}{data.frame}(spec)

\method{mona_parseSpec}{matrix}(spec)

\method{mona_parseSpec}{Spectrum2}(spec)
}
\arguments{
\item{spec}{A spectrum as either a character string of space deliminated
mz:intensity pairs or a dataframe with variables mz and intensity}
}
\value{
Provided a spectrum as a string returns a tibble and vice versa.
}
\description{
Interchange between mass spectrum formats for MONA queries and results
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Method for character strings

\item \code{data.frame}: Method for data frames

\item \code{matrix}: Method for matrix

\item \code{Spectrum2}: Method for Spectrum2 objects from MSnbase
}}

