% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_query_by_mol_identifier.R
\name{mona_query}
\alias{mona_query}
\title{Query Mass Bank of North America API by molecule identifier}
\usage{
mona_query(
  query = NULL,
  from = NULL,
  mass_tol_Da = NULL,
  ionization = NULL,
  ms_level = NULL,
  source_introduction = NULL
)
}
\arguments{
\item{query}{The query string}

\item{from}{The query type (text, name, InChIKey or molform)}

\item{mass_tol_Da}{Mass tolerance in Daltons for searching my molecular
weight.}

\item{ionization}{The ionization mode (positive, negative, both)}

\item{ms_level}{Which order of MS should be searched (all, MS1, MS2, MS3,
MS4).}

\item{source_introduction}{Which instrument platforms should be considered
(all, LC-MS, GC-MS, CE-MS)}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Submit a query to MONA API (https://mona.fiehnlab.ucdavis.edu) by text
string, molecular formula or (partial) InChI Key.
}
