% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_by_mol_identifier.R
\name{mona_query}
\alias{mona_query}
\title{Query Mass Bank of North America API by molecule identifier}
\usage{
mona_query(
  query,
  from = c("text", "name", "InChIKey", "molform", "mass"),
  mass_tol_Da = 0.1,
  ionization = c("positive", "negative"),
  ms_level = c("MS1", "MS2", "MS3", "MS4"),
  source_introduction = c("LC-MS", "GC-MS", "CE-MS")
)
}
\arguments{
\item{query}{The query string}

\item{from}{The query type (text, name, InChIKey or molform)}

\item{mass_tol_Da}{Mass tolerance in Daltons for searching my molecular
weight.}

\item{ionization}{The ionization mode (positive, negative, both)}

\item{ms_level}{Which order of MS should be searched (all, MS1, MS2, MS3,
MS4).}

\item{source_introduction}{Which instrument platforms should be considered
(all, LC-MS, GC-MS, CE-MS)}
}
\value{
A tibble containing the search results or NA if no matches.
}
\description{
Submit a query to MONA API (https://mona.fiehnlab.ucdavis.edu) by text
string, molecular formula or (partial) InChI Key.
}
