% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_by_spectrum.R
\name{mona_querySpec}
\alias{mona_querySpec}
\title{Query MONA API by a mass spectrum}
\usage{
mona_querySpec(
  spectrum = NULL,
  minSimilarity = NULL,
  precursorMZ = NULL,
  precursorToleranceDa = NULL,
  precursorTolerancePPM = NULL
)
}
\arguments{
\item{spectrum}{A spectrum as either a dataframe with variables mz,intensity or a space deliminated stinrg of mz:inensity pairs.}

\item{minSimilarity}{Minimum spectral similarity for searching}

\item{precursorMZ}{The precursor ion m/z}

\item{precursorToleranceDa}{Search tolerance in dalton}

\item{precursorTolerancePPM}{Search tolerance in ppm}
}
\value{
A tibble containing the search results of NA if an error occured.
}
\description{
Query MONA API by a mass spectrum
}
