% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_print_methods.R
\name{print.mona_id_query}
\alias{print.mona_id_query}
\title{print method for id query}
\usage{
\method{print}{mona_id_query}(x, ...)
}
\description{
print method for id query
}
\examples{
print(example_id_query)
}
