% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_print_methods.R
\name{print.mona_meta}
\alias{print.mona_meta}
\title{print method for meta data}
\usage{
\method{print}{mona_meta}(x, ...)
}
\arguments{
\item{x}{result of mona_getMeta, mona_getChem or mona_getSpec}

\item{...}{additional arguments}
}
\value{
Prints tibble info from a meta data query
}
\description{
print method for meta data
}
\examples{
print(mona_getMeta(example_id_query, 'category', 'mass spectrometry'))
}
