% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_print_methods.R
\name{print.mona_spec_query}
\alias{print.mona_spec_query}
\title{print method for spec query}
\usage{
\method{print}{mona_spec_query}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints tibble info from a spectrum query
}
\description{
print method for spec query
}
\examples{
print(example_spec_query)
}
